orders/stripe

+++ client/src/components/admin/AdminProductPreview.jsx
@@ -12,6 +12,10 @@ export default function AdminProductPreview({ product }) {
   const [deleteProduct] = useDeleteProductMutation();
 
   async function handleProductDelete() {
+    const confirmed = window.confirm(
+      'Are you sure you want to delete this item?'
+    );
+    if (!confirmed) return;
     try {
       await deleteProduct({ id: product._id });
       toast.success(`${product.name} succesfully deleted`);
